requirements.txt
asgiref==3.4.1
crispy==0.6.3
cycler==0.10.0
Django==3.2.8
django-crispy-forms==1.13.0
fabio==0.12.0
h5py==3.5.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pkg_resources==0.0.0
pyparsing==3.0.3
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.3
silx==0.15.2
six==1.16.0
sqlparse==0.4.2
